Refactored Pages out of App.js
