+++ src/App.css
@@ -2,12 +2,5 @@
   display: flex;
   padding-left: 10px;
   padding-right: 10px;
-  /* padding-top: 10px; */
   justify-content: center;
-}
-
-.Content {
-  width: 80vw;
-  margin-top: 10px;
-  margin-bottom: 10px;
 }
