+++ src/App.js
@@ -1,34 +1,25 @@
 import './App.css';
-import { useState, useEffect, useContext } from 'react';
+import { useState, useEffect } from 'react';
 import LoginService from './backend/services/login-service';
 import { UserTopArtistsProvider } from './backend/data/UserTopArtistsContext';
 import { UserTopTracksProvider } from './backend/data/UserTopTracksContext';
 import { MostRecentlyPlayedProvider } from './backend/data/MostRecentlyPlayedContext'
-import UserTopItemsContainer from './components/UserTopItems/UserTopItemsContainer';
-import { Box, Button } from '@mui/joy';
-import DevServiceContext from './backend/services/dev-service';
+import { Box } from '@mui/joy';
+import { DevServiceProvider } from './backend/services/dev-service-context';
+import LoginPage from './pages/LoginPage/LoginPage';
+import UserProfilePage from './pages/UserProfilePage/UserProfilePage';
 
 function App() {
-  const {updateIsDevMode, isDevMode} = useContext(DevServiceContext);
 
   const [spotifyAccessToken, setSpotifyAccessToken] = useState("");
 
   useEffect(() => {
     document.title = 'Spotify Analytics';
-    updateIsDevMode();
     if(window.location.hash) {
       const spotifyAccessParams = LoginService.getParamsAfterLoginRedirect();
       setSpotifyAccessToken(spotifyAccessParams.access_token);
     }
   }, [spotifyAccessToken])
-
-  const loginRedirect = () => {
-    if(isDevMode === "true"){
-      LoginService.loginRedirectDev();
-    } else {
-      LoginService.loginRedirectProd();
-    }
-  }
 
   if(spotifyAccessToken.length === 0){
     return (
@@ -37,14 +28,11 @@ function App() {
         className="App"
         sx={(theme) => ({
           backgroundColor: theme.variants.solid.neutral,
-          height: '100lvh',
-          justifyContent: "center",
-          alignItems: "center"
         })}
       >
-        <Box>
-          <Button size="lg" onClick={loginRedirect}>Login To Spotify</Button>
-        </Box>
+        <DevServiceProvider>
+          <LoginPage />
+        </DevServiceProvider>
       </Box>
     );
   }
@@ -57,16 +45,15 @@ function App() {
         backgroundColor: theme.variants.solid.neutral,
       })}
     >
-      <Box className='Content'>
-        <Button onClick={loginRedirect}>Re-login To Spotify</Button>
+      <DevServiceProvider>
         <UserTopArtistsProvider>
           <UserTopTracksProvider>
-            <MostRecentlyPlayedProvider>
-              <UserTopItemsContainer />
-            </MostRecentlyPlayedProvider>
+              <MostRecentlyPlayedProvider>
+                <UserProfilePage />
+              </MostRecentlyPlayedProvider>
           </UserTopTracksProvider>
         </UserTopArtistsProvider>
-      </Box>
+      </DevServiceProvider>
     </Box>
   );
 }
