+++ src/index.js
@@ -3,13 +3,10 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { DevServiceProvider } from './backend/services/dev-service';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-    <DevServiceProvider>
-        <App />
-    </DevServiceProvider>
+    <App />
 );
 
 // If you want to start measuring performance in your app, pass a function
