+++ src/pages/UserProfilePage/UserProfilePage.css
@@ -1,0 +1,11 @@
+.Content {
+    width: 80vw;
+    margin-top: 10px;
+    margin-bottom: 10px;
+    padding-left: 0px;
+    padding-right: 0px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+  }
+
